api/requirements/common.txt
Django~=3.1.1
djangorestframework
djangorestframework-jwt
Pillow
pylint
django-cors-headers
mysqlclient
boto3
django-storages
